fix error
deploy issue: dependencies
browser autofill not trigger onChange
add an util converting .env to azure app service format
text-to-speech not working in app service: 0x38 (SPXERR_AUDIO_SYS_LIBRARY_NOT_FOUND)

+++ backend/util/ops_junk_removal/blob_obj_remove.py
@@ -63,4 +63,7 @@ for result in search_client.search(search_text="*", select="id, imgPath"):
     if result["imgPath"] not in image_paths:
         del_items.append({"id": result["id"], "imgPath": result["imgPath"]})
 
-search_client.delete_documents(documents=del_items)
+if del_items:
+    search_client.delete_documents(documents=del_items)
+
+print('Done')
